refactor: Simplification in scikit-learn based estimators
