+++ syne_tune/optimizer/schedulers/searchers/bayesopt/sklearn/__init__.py
@@ -10,3 +10,14 @@
 # on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
+from syne_tune.optimizer.schedulers.searchers.bayesopt.sklearn.predictor import (  # noqa: F401
+    SKLearnPredictor,
+)
+from syne_tune.optimizer.schedulers.searchers.bayesopt.sklearn.estimator import (  # noqa: F401
+    SKLearnEstimator,
+)
+
+__all__ = [
+    "SKLearnPredictor",
+    "SKLearnEstimator",
+]
